+++ src/main/java/step4/Application.java
@@ -1,7 +1,8 @@
 package step4;
 
+import step4.controller.PlayGame;
 import step4.util.InputUtils;
-import step4.vehicle.Cars;
+import step4.domain.Cars;
 import step4.view.ResultView;
 
 public class Application {
