+++ src/main/java/step4/controller/PlayGame.java
@@ -1,15 +1,15 @@
-package step4;
+package step4.controller;
 
 import step4.util.RandomNumberGenerator;
-import step4.vehicle.Car;
-import step4.vehicle.Cars;
+import step4.domain.Car;
+import step4.domain.Cars;
 import step4.view.ResultView;
 
 public class PlayGame {
-    Cars cars;
-    ResultView resultView;
+    private final Cars cars;
+    private final ResultView resultView;
 
-    PlayGame(Cars cars, ResultView resultView) {
+    public PlayGame(Cars cars, ResultView resultView) {
         this.cars = cars;
         this.resultView = resultView;
     }
