+++ src/main/java/step4/domain/Cars.java
@@ -1,14 +1,13 @@
-package step4.vehicle;
+package step4.domain;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class Cars {
-    public static final int NUMBER_OF_PLAYERS = 2;
+    private static final int NUMBER_OF_PLAYERS = 2;
 
     private final List<Car> cars = new ArrayList<>();
-    private final List<Car> winners = new ArrayList<>();
 
     public Cars(String[] names) {
         if (names.length < NUMBER_OF_PLAYERS) {
