+++ src/main/java/step4/util/NumberGenerator.java
@@ -2,7 +2,7 @@ package step4.util;
 
 @FunctionalInterface
 public interface NumberGenerator {
-    int bound = 10;
+    int BOUND = 10;
 
     int getValue();
